+++ scripts/changelog.ts
@@ -192,7 +192,6 @@ async function exec(): Promise<void> {
         '**plugin-text**. Added list plugin',
         '**plugin-text**. Added colors plugin',
         '**plugin-text**. Added button to wrap in a blockquote plugin',
-        '**plugin-text**. Added Markdownshortcuts for Lists, Headings and Blockquotes.',
         '**plugin-rows**. Pass `remove` and `replace` functions to childs',
         '**all**. Added plugin icons and descriptions'
       ],
@@ -209,6 +208,11 @@ async function exec(): Promise<void> {
       internal: [
         '**demo**. Log errors only',
         '**demo**. Add additional examples for plugin-usage in Storybook'
+      ]
+    },
+    {
+      added: [
+        '**plugin-text**. Added Markdownshortcuts for Lists, Headings and Blockquotes.'
       ]
     }
   ])