+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased](https://github.com/edtr-io/edtr-io/compare/v0.5.0..HEAD)
+
+### Added
+
+- **plugin-text**. Added Markdownshortcuts for Lists, Headings and Blockquotes.
+
 ## [0.5.0](https://github.com/edtr-io/edtr-io/compare/v0.4.3..v0.5.0) - May 7, 2019
 
 ### Breaking Changes
@@ -24,7 +30,6 @@ All notable changes to this project will be documented in this file.
 - **plugin-text**. Added list plugin
 - **plugin-text**. Added colors plugin
 - **plugin-text**. Added button to wrap in a blockquote plugin
-- **plugin-text**. Added Markdownshortcuts for Lists, Headings and Blockquotes.
 - **plugin-rows**. Pass `remove` and `replace` functions to childs
 - **all**. Added plugin icons and descriptions